fix: move categorizeStatus/calculateMetrics to lib/jira-utils.ts
Next.js route files cannot export non-HTTP functions like categorizeStatus
and calculateMetrics. Moved them to src/lib/jira-utils.ts to satisfy
Next.js build requirements.
