+++ src/app/api/integrations/jira/route.ts
@@ -1,23 +1,12 @@
-import { getServerSession } from "next-auth";
 import { NextRequest } from "next/server";
+import { getServerSession } from "next-auth";
 import { authOptions } from "@/lib/auth";
 import { supabaseAdmin } from "@/lib/supabase";
 import { resolveAppUser, AppUser } from "@/lib/resolve-user";
 import { decryptToken } from "@/lib/crypto";
+import { JiraIssue, calculateMetrics } from "@/lib/jira-utils";
 
 export const dynamic = "force-dynamic";
-
-interface JiraIssue {
-  key: string;
-  summary: string;
-  status: string;
-  statusCategory: string;
-  created: string;
-  updated: string;
-  resolved: string | null;
-  assignee: string | null;
-  priority: string;
-}
 
 interface JiraCredentials {
   id: string;
@@ -98,46 +87,6 @@ async function fetchJiraIssues(
   }));
 }
 
-export function categorizeStatus(issue: JiraIssue): string {
-  if (issue.statusCategory === "done") {
-    return "Done";
-  }
-  if (issue.statusCategory === "indeterminate") {
-    return "In Progress";
-  }
-  return "To Do";
-}
-
-export function calculateMetrics(issues: JiraIssue[]) {
-  const toDo = issues.filter(
-    (i) => categorizeStatus(i) === "To Do"
-  ).length;
-  const inProgress = issues.filter(
-    (i) => categorizeStatus(i) === "In Progress"
-  ).length;
-  const done = issues.filter((i) => categorizeStatus(i) === "Done").length;
-
-  const resolvedIssues = issues.filter((i) => i.resolved !== null);
-  let avgTimeToClose: number | null = null;
-
-  if (resolvedIssues.length > 0) {
-    const totalHours = resolvedIssues.reduce((sum, issue) => {
-      const created = new Date(issue.created).getTime();
-      const resolved = new Date(issue.resolved!).getTime();
-      return sum + (resolved - created);
-    }, 0);
-    avgTimeToClose = Math.round(totalHours / resolvedIssues.length / 3600000);
-  }
-
-  return {
-    total: issues.length,
-    toDo,
-    inProgress,
-    done,
-    avgTimeToClose,
-  };
-}
-
 export async function GET(req: NextRequest) {
   const result = await requireUser();
   if ("error" in result) return result.error;
@@ -183,7 +132,7 @@ export async function GET(req: NextRequest) {
       metrics,
       recentIssues: issues.slice(0, 10),
     });
-  } catch (err) {
+  } catch {
     return Response.json(
       { error: "Failed to fetch Jira data" },
       { status: 502 }
